# api/app/routers/kundli.py
"""
Kundli router — Generate and retrieve Vedic birth charts.
Astrologer-only endpoints that use FreeAstroAPI for chart calculation.
"""
from fastapi import APIRouter, Depends, HTTPException
from sqlalchemy.orm import Session
from typing import List
from datetime import date as date_cls
from .. import models, schemas, database
from ..vedic_rishi_service import geocode_place
from ..free_astro_service import generate_full_kundli, generate_dasha_insights
from .auth import get_current_user
import asyncio

router = APIRouter(
    prefix="/kundli",
    tags=["Kundli"]
)


def _require_astrologer(current_user: models.User):
    """Helper to ensure the current user is an astrologer."""
    if current_user.role != models.UserRole.ASTROLOGER:
        raise HTTPException(status_code=403, detail="Only astrologers can access Kundli features")


@router.post("/generate", response_model=schemas.KundliReportResponse)
async def generate_kundli_report(
    request: schemas.KundliGenerateRequest,
    current_user: models.User = Depends(get_current_user),
    db: Session = Depends(database.get_db)
):
    """
    Generate a Kundli report for a seeker.
    Accepts birth details directly or a seeker_id to auto-fetch from profile.
    Geocodes place of birth and calls AstroAPI.
    Caches the result in kundli_reports table.
    """
    _require_astrologer(current_user)

    dob = request.date_of_birth
    tob = request.time_of_birth
    place = request.place_of_birth
    name = request.full_name or ""

    # If seeker_id provided, try to fetch details from profile
    if request.seeker_id:
        seeker_profile = db.query(models.SeekerProfile).filter(
            models.SeekerProfile.user_id == request.seeker_id
        ).first()
        if seeker_profile:
            if not dob and seeker_profile.date_of_birth:
                dob = seeker_profile.date_of_birth
            if not tob and seeker_profile.time_of_birth:
                tob = seeker_profile.time_of_birth
            if not place and seeker_profile.place_of_birth:
                place = seeker_profile.place_of_birth
            if not name and seeker_profile.full_name:
                name = seeker_profile.full_name

    # Validate required fields
    if not dob or not tob or not place:
        raise HTTPException(
            status_code=400,
            detail="date_of_birth, time_of_birth, and place_of_birth are required"
        )

    # Check for existing cached report with same birth details.
    # Reports generated before the switch to FreeAstroAPI carry the old
    # astroapi.dev response shape (no top-level "chart" key) and are
    # incompatible with the current frontend — treat those as stale and
    # regenerate rather than serving unusable cached data.
    existing = db.query(models.KundliReport).filter(
        models.KundliReport.date_of_birth == dob,
        models.KundliReport.time_of_birth == tob,
        models.KundliReport.place_of_birth == place,
    ).first()

    if existing and isinstance(existing.chart_data, dict) and "chart" in existing.chart_data:
        return existing

    # Geocode place of birth
    try:
        lat, lon = await geocode_place(place)
    except ValueError as e:
        raise HTTPException(status_code=400, detail=str(e))
    except Exception:
        raise HTTPException(status_code=500, detail="Geocoding service unavailable. Please try again.")

    # Call FreeAstroAPI for the full Kundli (chart, vargas, dasha, yogas, panchang, shadbala, ashtakavarga)
    try:
        chart_data = await generate_full_kundli(
            year=dob.year,
            month=dob.month,
            day=dob.day,
            hour=tob.hour,
            minute=tob.minute,
            latitude=lat,
            longitude=lon,
            timezone="Asia/Kolkata",
        )
    except ValueError as e:
        raise HTTPException(status_code=400, detail=str(e))
    except Exception as e:
        raise HTTPException(status_code=502, detail=f"FreeAstroAPI error: {str(e)}")

    # Save to database — refresh the stale row in place if one exists, otherwise insert new
    if existing:
        existing.full_name = name
        existing.latitude = lat
        existing.longitude = lon
        existing.chart_data = chart_data
        db.commit()
        db.refresh(existing)
        return existing

    report = models.KundliReport(
        seeker_id=request.seeker_id,
        generated_by=current_user.id,
        full_name=name,
        date_of_birth=dob,
        time_of_birth=tob,
        place_of_birth=place,
        latitude=lat,
        longitude=lon,
        timezone="Asia/Kolkata",
        chart_data=chart_data,
    )
    db.add(report)
    db.commit()
    db.refresh(report)

    return report


@router.get("/{report_id}/dasha-insights", response_model=schemas.KundliReportResponse)
async def get_dasha_insights(
    report_id: int,
    current_user: models.User = Depends(get_current_user),
    db: Session = Depends(database.get_db)
):
    """
    Get (and lazily compute) narrative Dasha Insights for a Kundli report.
    Cached on the report row per calendar day — repeat requests on the same
    day return the cached data without re-hitting FreeAstroAPI.
    """
    _require_astrologer(current_user)

    report = db.query(models.KundliReport).filter(
        models.KundliReport.id == report_id
    ).first()

    if not report:
        raise HTTPException(status_code=404, detail="Kundli report not found")

    today = date_cls.today()
    if report.dasha_insights_data and report.dasha_insights_date == today:
        return report

    if report.latitude is None or report.longitude is None:
        raise HTTPException(status_code=400, detail="Report is missing geocoded birth coordinates")

    try:
        insights = await generate_dasha_insights(
            year=report.date_of_birth.year,
            month=report.date_of_birth.month,
            day=report.date_of_birth.day,
            hour=report.time_of_birth.hour,
            minute=report.time_of_birth.minute,
            latitude=float(report.latitude),
            longitude=float(report.longitude),
            timezone=report.timezone or "Asia/Kolkata",
            reference_date=today.isoformat(),
        )
    except ValueError as e:
        raise HTTPException(status_code=400, detail=str(e))
    except Exception as e:
        raise HTTPException(status_code=502, detail=f"FreeAstroAPI error: {str(e)}")

    report.dasha_insights_data = insights
    report.dasha_insights_date = today
    db.commit()
    db.refresh(report)

    return report


@router.get("/{report_id}", response_model=schemas.KundliReportResponse)
def get_kundli_report(
    report_id: int,
    current_user: models.User = Depends(get_current_user),
    db: Session = Depends(database.get_db)
):
    """Get a specific Kundli report by ID."""
    _require_astrologer(current_user)

    report = db.query(models.KundliReport).filter(
        models.KundliReport.id == report_id
    ).first()

    if not report:
        raise HTTPException(status_code=404, detail="Kundli report not found")

    return report


@router.get("/seeker/{seeker_id}", response_model=List[schemas.KundliReportResponse])
def get_seeker_kundli_reports(
    seeker_id: int,
    current_user: models.User = Depends(get_current_user),
    db: Session = Depends(database.get_db)
):
    """Get all Kundli reports for a specific seeker."""
    _require_astrologer(current_user)

    reports = db.query(models.KundliReport).filter(
        models.KundliReport.seeker_id == seeker_id
    ).order_by(models.KundliReport.created_at.desc()).all()

    return reports


@router.get("/history/all", response_model=List[schemas.KundliReportResponse])
def get_astrologer_kundli_history(
    current_user: models.User = Depends(get_current_user),
    db: Session = Depends(database.get_db)
):
    """Get all Kundli reports generated by the current astrologer."""
    _require_astrologer(current_user)

    reports = db.query(models.KundliReport).filter(
        models.KundliReport.generated_by == current_user.id
    ).order_by(models.KundliReport.created_at.desc()).all()

    return reports