# api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KundliGenerateRequest": {
        "properties": {
          "date_of_birth": {
            "format": "date",
            "title": "Date Of Birth",
            "type": "string"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "place_of_birth": {
            "title": "Place Of Birth",
            "type": "string"
          },
          "seeker_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seeker Id"
          },
          "time_of_birth": {
            "format": "time",
            "title": "Time Of Birth",
            "type": "string"
          }
        },
        "required": [
          "date_of_birth",
          "time_of_birth",
          "place_of_birth"
        ],
        "title": "KundliGenerateRequest",
        "type": "object"
      },
      "KundliReportResponse": {
        "properties": {
          "chart_data": {
            "additionalProperties": true,
            "title": "Chart Data",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dasha_insights_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dasha Insights Data"
          },
          "date_of_birth": {
            "format": "date",
            "title": "Date Of Birth",
            "type": "string"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "place_of_birth": {
            "title": "Place Of Birth",
            "type": "string"
          },
          "seeker_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seeker Id"
          },
          "time_of_birth": {
            "format": "time",
            "title": "Time Of Birth",
            "type": "string"
          }
        },
        "required": [
          "id",
          "seeker_id",
          "full_name",
          "date_of_birth",
          "time_of_birth",
          "place_of_birth",
          "latitude",
          "longitude",
          "chart_data",
          "created_at"
        ],
        "title": "KundliReportResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Aadikarta API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/kundli/generate": {
      "post": {
        "description": "Generate a Kundli report for a seeker.\nAccepts birth details directly or a seeker_id to auto-fetch from profile.\nGeocodes place of birth and calls AstroAPI.\nCaches the result in kundli_reports table.",
        "operationId": "generate_kundli_report_kundli_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KundliGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KundliReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Generate Kundli Report",
        "tags": [
          "Kundli"
        ]
      }
    },
    "/kundli/history/all": {
      "get": {
        "description": "Get all Kundli reports generated by the current astrologer.",
        "operationId": "get_astrologer_kundli_history_kundli_history_all_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KundliReportResponse"
                  },
                  "title": "Response Get Astrologer Kundli History Kundli History All Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Astrologer Kundli History",
        "tags": [
          "Kundli"
        ]
      }
    },
    "/kundli/seeker/{seeker_id}": {
      "get": {
        "description": "Get all Kundli reports for a specific seeker.",
        "operationId": "get_seeker_kundli_reports_kundli_seeker__seeker_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "seeker_id",
            "required": true,
            "schema": {
              "title": "Seeker Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KundliReportResponse"
                  },
                  "title": "Response Get Seeker Kundli Reports Kundli Seeker  Seeker Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Seeker Kundli Reports",
        "tags": [
          "Kundli"
        ]
      }
    },
    "/kundli/{report_id}": {
      "get": {
        "description": "Get a specific Kundli report by ID.",
        "operationId": "get_kundli_report_kundli__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KundliReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Kundli Report",
        "tags": [
          "Kundli"
        ]
      }
    },
    "/kundli/{report_id}/dasha-insights": {
      "get": {
        "description": "Get (and lazily compute) narrative Dasha Insights for a Kundli report.\nCached on the report row per calendar day \u2014 repeat requests on the same\nday return the cached data without re-hitting FreeAstroAPI.",
        "operationId": "get_dasha_insights_kundli__report_id__dasha_insights_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KundliReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Dasha Insights",
        "tags": [
          "Kundli"
        ]
      }
    },
...
  }
}